refactor package of ParameterRewriter
