+++ sharding-core/sharding-core-rewrite/src/main/java/org/apache/shardingsphere/core/rewrite/SQLRewriteEngine.java
@@ -28,8 +28,8 @@ import org.apache.shardingsphere.core.rewrite.constant.EncryptDerivedColumnType;
 import org.apache.shardingsphere.core.rewrite.constant.ShardingDerivedColumnType;
 import org.apache.shardingsphere.core.rewrite.parameter.builder.ParameterBuilder;
 import org.apache.shardingsphere.core.rewrite.parameter.builder.ParameterBuilderFactory;
-import org.apache.shardingsphere.core.rewrite.parameter.rewriter.EncryptParameterBuilderFactory;
-import org.apache.shardingsphere.core.rewrite.parameter.rewriter.ShardingParameterBuilderFactory;
+import org.apache.shardingsphere.core.rewrite.parameter.rewriter.encrypt.EncryptParameterBuilderFactory;
+import org.apache.shardingsphere.core.rewrite.parameter.rewriter.sharding.ShardingParameterBuilderFactory;
 import org.apache.shardingsphere.core.rewrite.token.SQLTokenGenerators;
 import org.apache.shardingsphere.core.rewrite.token.builder.BaseTokenGeneratorBuilder;
 import org.apache.shardingsphere.core.rewrite.token.builder.EncryptTokenGenerateBuilder;
